fix: cleanup dispatching contextMenu events
